// src/enums/product.ts
/* eslint-disable no-unused-vars */
// TODO: Fix eslint

export enum ProductTier {
  Basic = 'Basic',
  Premium = 'Premium',
  Deluxe = 'Deluxe',
}

export enum ProductTheme {
  Dark = 'Dark',
  Light = 'Light',
  Colorful = 'Colorful',
  Halloween = 'Halloween',
}

export enum ProductCategory {
  All = '',
  UpperBody = 'Upper Body',
  LowerBody = 'Lower Body',
  Hat = 'Hat',
  Shoes = 'Shoes',
  Accessory = 'Accessory',
  Legendary = 'Legendary',
  Mythic = 'Mythic',
  Epic = 'Epic',
  Rare = 'Rare',
}